backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
drf-yasg
flake8
djangorestframework-simplejwt==5.3.1
Pillow
psycopg2-binary
django-cors-headers
elasticsearch-dsl
django-elasticsearch-dsl
django-redis
